+++ jpa/plugins/org.eclipse.jpt.core/src/org/eclipse/jpt/core/internal/context/orm/GenericOrmDiscriminatorColumn.java
@@ -70,11 +70,11 @@ public class GenericOrmDiscriminatorColumn extends AbstractOrmNamedColumn<XmlDis
 		firePropertyChanged(DiscriminatorColumn.SPECIFIED_DISCRIMINATOR_TYPE_PROPERTY, oldDiscriminatorType, newSpecifiedDiscriminatorType);
 	}
 		
-	public Integer getLength() {
-		return (this.getSpecifiedLength() == null) ? this.getDefaultLength() : this.getSpecifiedLength();
+	public int getLength() {
+		return (this.getSpecifiedLength() == null) ? this.getDefaultLength() : this.getSpecifiedLength().intValue();
 	}
 
-	public Integer getDefaultLength() {
+	public int getDefaultLength() {
 		return DiscriminatorColumn.DEFAULT_LENGTH;
 	}
